fix: remove dead references to DB features in `sqlx-core`
This reactivates a couple of tests that had stopped working.
