+++ sqlx-core/src/query_builder.rs
@@ -593,118 +593,3 @@ where
         self
     }
 }
-
-#[cfg(all(test, feature = "postgres"))]
-mod test {
-    use crate::postgres::Postgres;
-
-    use super::*;
-
-    #[test]
-    fn test_new() {
-        let qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("SELECT * FROM users");
-        assert_eq!(qb.query, "SELECT * FROM users");
-    }
-
-    #[test]
-    fn test_push() {
-        let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("SELECT * FROM users");
-        let second_line = " WHERE last_name LIKE '[A-N]%;";
-        qb.push(second_line);
-
-        assert_eq!(
-            qb.query,
-            "SELECT * FROM users WHERE last_name LIKE '[A-N]%;".to_string(),
-        );
-    }
-
-    #[test]
-    #[should_panic]
-    fn test_push_panics_when_no_arguments() {
-        let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("SELECT * FROM users;");
-        qb.arguments = None;
-
-        qb.push("SELECT * FROM users;");
-    }
-
-    #[test]
-    fn test_push_bind() {
-        let mut qb: QueryBuilder<'_, Postgres> =
-            QueryBuilder::new("SELECT * FROM users WHERE id = ");
-
-        qb.push_bind(42i32)
-            .push(" OR membership_level = ")
-            .push_bind(3i32);
-
-        assert_eq!(
-            qb.query,
-            "SELECT * FROM users WHERE id = $1 OR membership_level = $2"
-        );
-    }
-
-    #[test]
-    fn test_build() {
-        let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("SELECT * FROM users");
-
-        qb.push(" WHERE id = ").push_bind(42i32);
-        let query = qb.build();
-
-        assert_eq!(
-            query.statement.unwrap_left(),
-            "SELECT * FROM users WHERE id = $1"
-        );
-        assert_eq!(query.persistent, true);
-    }
-
-    #[test]
-    fn test_reset() {
-        let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("");
-
-        let _query = qb
-            .push("SELECT * FROM users WHERE id = ")
-            .push_bind(42i32)
-            .build();
-
-        qb.reset();
-
-        assert_eq!(qb.query, "");
-    }
-
-    #[test]
-    fn test_query_builder_reuse() {
-        let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("");
-
-        let _query = qb
-            .push("SELECT * FROM users WHERE id = ")
-            .push_bind(42i32)
-            .build();
-
-        qb.reset();
-
-        let query = qb.push("SELECT * FROM users WHERE id = 99").build();
-
-        assert_eq!(
-            query.statement.unwrap_left(),
-            "SELECT * FROM users WHERE id = 99"
-        );
-    }
-
-    #[test]
-    fn test_query_builder_with_args() {
-        let mut qb: QueryBuilder<'_, Postgres> = QueryBuilder::new("");
-
-        let query = qb
-            .push("SELECT * FROM users WHERE id = ")
-            .push_bind(42i32)
-            .build();
-
-        let mut qb: QueryBuilder<'_, Postgres> =
-            QueryBuilder::new_with(query.sql(), query.take_arguments());
-        let query = qb.push("OR membership_level =").push_bind(3i32).build();
-
-        assert_eq!(
-            query.sql(),
-            "SELECT * FROM users WHERE id = $1 OR membership_level = $2"
-        );
-    }
-}
